# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ERP API Documentation",
        "version": "1.0"
    },
    "paths": {
        "/api/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "List all permissions",
                "operationId": "dc48327de70639700359c6735af4e190",
                "responses": {
                    "200": {
                        "description": "List of permissions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "permissions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "edit users"
                                                    },
                                                    "guard_name": {
                                                        "type": "string",
                                                        "example": "api"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Create a new permission",
                "operationId": "3f0ba02c9960ce5e3d6694c942dfe958",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "edit users"
                                    },
                                    "guard_name": {
                                        "type": "string",
                                        "example": "api"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission created successfully"
                    }
                }
            }
        },
        "/api/permissions/{id}": {
            "put": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Update a permission",
                "operationId": "2c8722483817e4c4baf7d5ce241efdf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "edit articles"
                                    },
                                    "guard_name": {
                                        "type": "string",
                                        "example": "api"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Delete a permission",
                "operationId": "f5b541734892974e0e028445ad4ecfdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission deleted successfully"
                    }
                }
            }
        },
...
    },
...
}